Update eval.js
uses a class to create a singleton with a private value, the module. This way the wrapper functions have access to the module, but not anyone else. That way, everything will be parsed as needed.

The parse function splits apart a function into its parameters, and the body. the function is toString'd, which means it should be pretty much exactly how it was typed out.

arguments passed can't be functions, since the args are handled by socketlib, but if you write a function that sends a string version of a function over and save it, then you can always just refer to it. Ie, memoize the function locally on other computers, and then use that.